+++ pkg/function/wait_for_snapshot_completion.go
@@ -65,7 +65,7 @@ func waitForSnapshotsCompletion(ctx context.Context, snapshotinfo string, profil
 	}
 	for _, pvcInfo := range PVCData {
 		config := make(map[string]string)
-		if err = ValidateProfile(profile, pvcInfo.Type); err != nil {
+		if err = ValidateLocationForBlockstorage(profile, pvcInfo.Type); err != nil {
 			return errors.Wrap(err, "Profile validation failed")
 		}
 		switch pvcInfo.Type {
